+++ src/validator/constraints/other/compound.integration.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { createValidator } from '../../validator-factory/create-validator';
+import { createValidator } from '@/validator/factory/create-validator';
 import { notBlank } from '../basic/not-blank';
 import { email } from '../string/email';
 import { compound } from './compound';
